helm: add conftest-test to check-helm-tests target
Reduce the number of make targets that contributors need to be aware of.
Fold conftest-test into check-helm-tests target.

Add a script to check if helm dependency update needs to be run and
skip it if it does not.

Reuse the generated files for check-helm-tests in conftest by not throwing
away the intermediate results.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>